Save make_player_key as separate script
