+++ python/make_player_key.py
@@ -1,0 +1,11 @@
+import unicodedata
+import re
+
+# Clean a player's name and generate a unique key for consistent cross-site merging
+def make_player_key(name):
+    normalized_name = unicodedata.normalize("NFD", name).encode("ascii", "ignore").decode("utf-8")  # Remove accents
+    cleaned_name = normalized_name.lower().strip()  # Convert to lowercase and trim spaces
+    cleaned_name = re.sub(r"\s+", "-", cleaned_name)  # Replace spaces with hyphens
+    cleaned_name = re.sub(r"[^\w-]", "", cleaned_name)  # Remove non-alphanumeric characters
+    player_key = re.sub(r"-(sr|jr|ii|iii|iv|v|vi|vii)$", "", cleaned_name)  # Remove common suffixes
+    return player_key
